Change training data to hold byte[] instead of double[]

+++ Artififical Neural Network/Neural Network/Training/TrainingData.cs
@@ -2,10 +2,10 @@
 {
     public class TrainingData
     {
-        public double[] inputData;
-        public int correctOutputNeuron;
+        public byte[] inputData;
+        public byte correctOutputNeuron;
 
-        public TrainingData(double[] input, int correctOuput)
+        public TrainingData(byte[] input, byte correctOuput)
         {
             inputData = input;
             correctOutputNeuron = correctOuput;
